% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_words}
\alias{count_words}
\title{Function 3: counting words}
\usage{
count_words(text)
}
\arguments{
\item{text}{A text source}
}
\value{
A number of how many words are in the text.
}
\description{
Function 3: counting words
}
