% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_stopwords}
\alias{remove_stopwords}
\title{Function 2: remove_stopwords}
\usage{
remove_stopwords(text)
}
\arguments{
\item{text}{A text source}
}
\value{
removes stopwords from the text
}
\description{
Function 2: remove_stopwords
}
\examples{
Hello! This is the example text.
Output: Hello! This example text.

}
